+++ src/domain/User.kt
@@ -1,1 +1,3 @@
+package domain
+
 data class User(val login: String, val hash: String, val salt: String)
