+++ src/domain/UserSession.kt
@@ -1,3 +1,5 @@
+package domain
+
 import java.util.*
 
 data class UserSession(
