+++ src/domain/UsersResources.kt
@@ -1,1 +1,5 @@
+package domain
+
+import Role
+
 data class UsersResources(val path: String?, val role: Role, val login: String)
